# a7_openapi.yaml
openapi: 3.0.0
info:
 version: '1.0'
 title: 'Feup Times'
 description: 'Web Resources Specification (A7) for Feup Times'
servers:
- url: http://lbaw.fe.up.pt
  description: Production server
...
paths:
...
 /welcome:
  get:
    operationId: R201
    summary: 'R201: View main page'
    description: 'Show the main page. Access: PUB'
    tags:
      - 'M02: Posts'
    responses:
      '200':
        description: 'Ok. Show [UI01](https://git.fe.up.pt/lbaw/lbaw2324/lbaw2331/-/wikis/uploads/d5e82533d57061229eb6945830db73fb/image.png)'
      '400':
        description: 'Bad Request.'
 /welcome/post/{id_recentNews}:
  get:
    operationId: R202
    summary: 'R202: View recent news'
    description: 'Show the most recent news. Access: PUB'
    tags:
      - 'M02: Posts'
    parameters:
        - in: path
          name: id_recentNews
          schema:
            type: string
          description: 'Specifies the number of the recentNews id for the query'
          required: true
    responses:
      '200':
        description: 'Ok. Show the most recent news.'
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Post'
      '400':
        description: 'Bad Request.'
...
 /welcome/post/{id_topNews}:
  get:
    operationId: R204
    summary: 'R204: View top news'
    description: 'Show the news with the popularity. Access: PUB'
    tags:
      - 'M02: Posts'
    parameters:
      - in: path
        name: id_topNews
        schema:
          type: string
        description: 'Specifies the number of the topNews id for the query'
        required: true
    responses:
      '200':
        description: 'Show the news with the most popularity.'
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Post'
      '400':
        description: 'Bad Request.'
...
 /welcome/post/{id_post}:
  post:
    operationId: R209
    summary: 'R209: Upvote or Downvote a post'
    description: 'Allows a member to rate a post. Access: USR'
    tags:
      - 'M02: Posts'
    parameters:
      - in: path
        name: id_post
        schema:
          type: integer
        description: 'The id of the post for the query.'
        required: true
    requestBody:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Vote'
    responses:
      '200':
        description: 'Rated post successfully.'
      '400':
        description: 'Bad Request.'
      '403':
        description: 'Forbidden access.'
      '404':
        description: 'Post not found.'
...
 /welcome/post/{id_post1}:
  delete:
    operationId: R211
    summary: 'R211: Delete a post'
    description: 'Deletes a post. Access: OWN, ADM'
    tags:
      - 'M02: Posts'
    parameters:
    - in: path
      name: id_post1
      schema:
        type: integer
      description: 'The id of the post to be deleted.'
      required: true
    responses:
      '302':
        description: 'Redirect after updating the post content.'
        headers:
          Location:
            schema:
              type: string
            examples:
              302Success:
                description: 'Successfulty deleted. Redirect to previous page.'
              302Error:
                description: 'Failed to delete. Redirect to previous page.'
  patch:
    operationId: R212
    summary: 'R212: Edit a post'
    description: 'Edits a post. Access: OWN'
    tags:
      - 'M02: Posts'
    parameters:
    - in: path
      name: id_post1
      schema:
        type: integer
      description: 'The id of the post to be deleted.'
      required: true
    responses:
      '200':
        description: 'Post edited successfully.'
      '400':
        description: 'Bad Request.'
      '403':
        description: 'Forbidden access.'
      '404':
        description: 'Post not found.'
  post:
    operationId: R213
    summary: 'R213: Insert a topic to a post'
    description: 'Inserts a topic to a post. Access: OWN'
    tags:
      - 'M02: Posts'
    parameters:
    - in: path
      name: id_post1
      schema:
        type: integer
      description: 'The id of the post to be deleted.'
      required: true
    requestBody:
      required: true
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Topic'
    responses:
      '200':
        description: 'Topic inserted successfully.'
      '400':
        description: 'Bad Request.'
      '403':
        description: 'Forbidden access.'
      '404':
        description: 'Post not found.'


# M03: Members
...
components:
  schemas:
    Topic:
        type: object
        properties:
          id:
            type: integer
          title:
            type: string
          description:
            type: string
...
    Post:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date_time:
          type: string
          format: date-time
        popularity:
          type: integer
        id_user:
          type: integer
        id_topic:
          type: integer
    Vote:
      type: object
      properties:
        id:
          type: integer
        rating:
          type: integer
        id_post:
          type: integer
        id_user:
          type: integer
...